requirements.txt
streamlit
aiohttp
beautifulsoup4
python-dotenv
fpdf
serpapi
requests
